Cleanup parser code
