+++ src/parser/parser.ts
@@ -6,7 +6,7 @@ import { SourCParser2 } from '../lang/SourCParser2'
 import { Context, ErrorSeverity } from '../types'
 import { ProgramContext } from './../lang/SourCParser2'
 import { FatalSyntaxError } from './error'
-import { contextToLocation } from './parser.util'
+import { contextToLocation } from './utils'
 import { Visitor } from './visitor'
 
 function convertSource(source: string, ctx: ProgramContext): es.Program {
@@ -52,122 +52,3 @@ export function parse(source: string, context: Context) {
     return undefined
   }
 }
-
-/*
-import { CalcLexer } from '../lang/CalcLexer'
-import {
-  AdditionContext,
-  CalcParser,
-  DivisionContext,
-  ExpressionContext,
-  MultiplicationContext,
-  NumberContext,
-  ParenthesesContext,
-  PowerContext,
-  StartContext,
-  SubtractionContext
-} from '../lang/CalcParser'
-import { CalcVisitor } from '../lang/CalcVisitor'
-
-class ExpressionGenerator implements CalcVisitor<es.Expression> {
-  visitNumber(ctx: NumberContext): es.Expression {
-    return {
-      type: 'Literal',
-      value: parseInt(ctx.text),
-      raw: ctx.text,
-      loc: contextToLocation(ctx)
-    }
-  }
-  visitParentheses(ctx: ParenthesesContext): es.Expression {
-    return this.visit(ctx.expression())
-  }
-  visitPower(ctx: PowerContext): es.Expression {
-    return {
-      type: 'BinaryExpression',
-      operator: '^',
-      left: this.visit(ctx._left),
-      right: this.visit(ctx._right),
-      loc: contextToLocation(ctx)
-    }
-  }
-
-  visitMultiplication(ctx: MultiplicationContext): es.Expression {
-    return {
-      type: 'BinaryExpression',
-      operator: '*',
-      left: this.visit(ctx._left),
-      right: this.visit(ctx._right),
-      loc: contextToLocation(ctx)
-    }
-  }
-  visitDivision(ctx: DivisionContext): es.Expression {
-    return {
-      type: 'BinaryExpression',
-      operator: '/',
-      left: this.visit(ctx._left),
-      right: this.visit(ctx._right),
-      loc: contextToLocation(ctx)
-    }
-  }
-  visitAddition(ctx: AdditionContext): es.Expression {
-    return {
-      type: 'BinaryExpression',
-      operator: '+',
-      left: this.visit(ctx._left),
-      right: this.visit(ctx._right),
-      loc: contextToLocation(ctx)
-    }
-  }
-
-  visitSubtraction(ctx: SubtractionContext): es.Expression {
-    return {
-      type: 'BinaryExpression',
-      operator: '-',
-      left: this.visit(ctx._left),
-      right: this.visit(ctx._right),
-      loc: contextToLocation(ctx)
-    }
-  }
-
-  visitExpression?: ((ctx: ExpressionContext) => es.Expression) | undefined
-  visitStart?: ((ctx: StartContext) => es.Expression) | undefined
-
-  visit(tree: ParseTree): es.Expression {
-    return tree.accept(this)
-  }
-  visitChildren(node: RuleNode): es.Expression {
-    const expressions: es.Expression[] = []
-    for (let i = 0; i < node.childCount; i++) {
-      expressions.push(node.getChild(i).accept(this))
-    }
-    return {
-      type: 'SequenceExpression',
-      expressions
-    }
-  }
-  visitTerminal(node: TerminalNode): es.Expression {
-    return node.accept(this)
-  }
-
-  visitErrorNode(node: ErrorNode): es.Expression {
-    throw new FatalSyntaxError(
-      {
-        start: {
-          line: node.symbol.line,
-          column: node.symbol.charPositionInLine
-        },
-        end: {
-          line: node.symbol.line,
-          column: node.symbol.charPositionInLine + 1
-        }
-      },
-      `invalid syntax ${node.text}`
-    )
-  }
-}
-
-function convertExpression(expression: ExpressionContext): es.Expression {
-  const generator = new ExpressionGenerator()
-  return expression.accept(generator)
-}
-*/
