+++ src/parser/visitor.ts
@@ -72,7 +72,7 @@ import {
   getUnaryOp,
   getUpdateOp,
   nodeToLocation
-} from './parser.util'
+} from './utils'
 
 export class Visitor implements SourCParser2Visitor<es.Node> {
   private PARSER_MISCONFIG = 'Parser is misconfigured'
@@ -596,8 +596,8 @@ export class Visitor implements SourCParser2Visitor<es.Node> {
     const updateNode = ctx._incrExpr
       ? (this.visit(ctx._incrExpr) as es.Expression)
       : ctx._incrAssgn
-      ? (this.visit(ctx._incrAssgn) as es.AssignmentExpression)
-      : undefined
+        ? (this.visit(ctx._incrAssgn) as es.AssignmentExpression)
+        : undefined
 
     return {
       ...contextToLocation(ctx),
